improve getrests.py

+++ scripts/getrests.py
@@ -4,11 +4,15 @@ import os
 import shutil
 srcpath="/proj/CESM2/archive/"
 destpath="/proj/jedwards/restarts/"
-dryrun = True
+dryrun = False
 
 for restyr in glob.iglob(srcpath+"*/rest/*"):
     yr = int(os.path.basename(restyr)[0:4])
     mon = int(os.path.basename(restyr)[5:7])
+    newdir = restyr.replace(srcpath,destpath)
+    if os.path.isdir(newdir):
+        print("Skipping {}, directory exists".format(newdir))
+        continue
     if (yr >= 1850 and yr <= 1870) or (yr >= 2010 and yr <= 2015):
         if mon != 1:
             # first delete unneeded files, then move directory
@@ -17,15 +21,14 @@ for restyr in glob.iglob(srcpath+"*/rest/*"):
                 if not dryrun:
                     os.remove(_file)
             print("moving dir {}".format(restyr))
-            if not dryrun:
+            if not dryrun and not os.path.isdir(newdir):
                 try:
                     shutil.move(restyr, restyr.replace(srcpath,destpath))
                 except OSError as e:
                     print("ERROR: {} - {}.".format(e.filename, e.strerror))
         else:
-            newdir = restyr.replace(srcpath,destpath)
             print("create dir {}".format(newdir))
-            if not dryrun:
+            if not dryrun and not os.path.isdir(newdir):
                 os.makedirs(newdir)
             for _file in glob.iglob(restyr+"/*"):
                 if ".h" not in _file and ".rh" not in _file:
